Fixing acl logging, moving challenge code out of limit
