+++ src/CMakeLists.txt
@@ -162,6 +162,7 @@ SET(SRCS
     core/scopes.cc
     core/scopes_configs.cc
     core/rules.cc
+    core/challenge.cc
     # ------------------------------------------------------
     # parsers
     # ------------------------------------------------------
@@ -191,7 +192,6 @@ list(APPEND SRCS
     # ------------------------------------------------------
     # limit
     # ------------------------------------------------------
-    limit/challenge.cc
     limit/rl_obj.cc
     limit/enforcer.cc
     limit/config.cc
